-Use double for all size parameters.  - Fix a crash bug in View when timer slot is NULL.  - Fix a potential draw issue in View when dealing with popup window    optimization.  - Fix window position issue in SingleViewHost.  - Add BeginResizeDrag() and BeginMoveDrag() method to ViewHostInterface.

+++ extensions/gtk_edit_element/gtk_edit_element.cc
@@ -151,11 +151,11 @@ void GtkEditElement::SetPasswordChar(const char *passwordChar) {
   impl_->SetPasswordChar(passwordChar);
 }
 
-int GtkEditElement::GetSize() const {
+double GtkEditElement::GetSize() const {
   return impl_->GetFontSize();
 }
 
-void GtkEditElement::SetSize(int size) {
+void GtkEditElement::SetSize(double size) {
   impl_->SetFontSize(size);
 }
