+++ ggadget/gtk/pixbuf_image.h
@@ -35,8 +35,8 @@ class PixbufImage : public CairoImageBase {
 
  public:
   virtual CanvasInterface *GetCanvas() const;
-  virtual size_t GetWidth() const;
-  virtual size_t GetHeight() const;
+  virtual double GetWidth() const;
+  virtual double GetHeight() const;
   virtual bool IsFullyOpaque() const;
 
  private:
